adapt to ocrd v3 (process→process_page_pcgts)…
- separate `setup` / `_init` from processing
  - inherit/override `_init`
  - init Tesseract API once, keep as class attribute
- define `process_page_pcgts` instead of `process`:
  - no input file loop
  - no METS action - just return tuples for derived images
- move shared utility functions into `common` module
- generate etree mapping by exporting PAGE model
