+++ ocrd_tesserocr/common.py
@@ -1,0 +1,396 @@
+import itertools
+from PIL import Image, ImageStat
+
+import numpy as np
+from scipy.sparse.csgraph import minimum_spanning_tree
+from shapely.geometry import Polygon, LineString
+from shapely.ops import unary_union, nearest_points, orient
+from shapely import set_precision
+
+
+from ocrd_utils import (
+    getLogger,
+    polygon_from_points,
+    points_from_polygon,
+)
+from ocrd_models.ocrd_page import (
+    ReadingOrderType,
+    RegionRefType,
+    RegionRefIndexedType,
+    OrderedGroupType,
+    OrderedGroupIndexedType,
+    UnorderedGroupType,
+    UnorderedGroupIndexedType,
+    PageType,
+    TextEquivType,
+)
+from ocrd_models.ocrd_page_generateds import (
+    ReadingDirectionSimpleType,
+    TextLineOrderSimpleType,
+)
+
+
+def page_element_unicode0(element):
+    """Get Unicode string of the first text result."""
+    if element.get_TextEquiv():
+        return element.get_TextEquiv()[0].Unicode or ''
+    else:
+        return ''
+
+def page_element_conf0(element):
+    """Get confidence (as float value) of the first text result."""
+    if element.get_TextEquiv():
+        # generateDS does not convert simpleType for attributes (yet?)
+        return float(element.get_TextEquiv()[0].conf or "1.0")
+    return 1.0
+
+def page_get_reading_order(ro, rogroup):
+    """Add all elements from the given reading order group to the given dictionary.
+    
+    Given a dict ``ro`` from layout element IDs to ReadingOrder element objects,
+    and an object ``rogroup`` with additional ReadingOrder element objects,
+    add all references to the dict, traversing the group recursively.
+    """
+    regionrefs = list()
+    if isinstance(rogroup, (OrderedGroupType, OrderedGroupIndexedType)):
+        regionrefs = (rogroup.get_RegionRefIndexed() +
+                      rogroup.get_OrderedGroupIndexed() +
+                      rogroup.get_UnorderedGroupIndexed())
+    if isinstance(rogroup, (UnorderedGroupType, UnorderedGroupIndexedType)):
+        regionrefs = (rogroup.get_RegionRef() +
+                      rogroup.get_OrderedGroup() +
+                      rogroup.get_UnorderedGroup())
+    for elem in regionrefs:
+        ro[elem.get_regionRef()] = elem
+        if not isinstance(elem, (RegionRefType, RegionRefIndexedType)):
+            page_get_reading_order(ro, elem)
+        
+def page_update_higher_textequiv_levels(level, pcgts, overwrite=True):
+    """Update the TextEquivs of all PAGE-XML hierarchy levels above ``level`` for consistency.
+    
+    Starting with the lowest hierarchy level chosen for processing,
+    join all first TextEquiv.Unicode (by the rules governing the respective level)
+    into TextEquiv.Unicode of the next higher level, replacing them.
+    If ``overwrite`` is false and the higher level already has text, keep it.
+    
+    When two successive elements appear in a ``Relation`` of type ``join``,
+    then join them directly (without their respective white space).
+    
+    Likewise, average all first TextEquiv.conf into TextEquiv.conf of the next higher level.
+    
+    In the process, traverse the words and lines in their respective ``readingDirection``,
+    the (text) regions which contain lines in their respective ``textLineOrder``, and
+    the (text) regions which contain text regions in their ``ReadingOrder``
+    (if they appear there as an ``OrderedGroup``).
+    Where no direction/order can be found, use XML ordering.
+    
+    Follow regions recursively, but make sure to traverse them in a depth-first strategy.
+    """
+    page = pcgts.get_Page()
+    relations = page.get_Relations() # get RelationsType
+    if relations:
+        relations = relations.get_Relation() # get list of RelationType
+    else:
+        relations = []
+    joins = list() # 
+    for relation in relations:
+        if relation.get_type() == 'join': # ignore 'link' type here
+            joins.append((relation.get_SourceRegionRef().get_regionRef(),
+                          relation.get_TargetRegionRef().get_regionRef()))
+    reading_order = dict()
+    ro = page.get_ReadingOrder()
+    if ro:
+        page_get_reading_order(reading_order, ro.get_OrderedGroup() or ro.get_UnorderedGroup())
+    if level != 'region':
+        for region in page.get_AllRegions(classes=['Text']):
+            # order is important here, because regions can be recursive,
+            # and we want to concatenate by depth first;
+            # typical recursion structures would be:
+            #  - TextRegion/@type=paragraph inside TextRegion
+            #  - TextRegion/@type=drop-capital followed by TextRegion/@type=paragraph inside TextRegion
+            #  - any region (including TableRegion or TextRegion) inside a TextRegion/@type=footnote
+            #  - TextRegion inside TableRegion
+            subregions = region.get_TextRegion()
+            if subregions: # already visited in earlier iterations
+                # do we have a reading order for these?
+                # TODO: what if at least some of the subregions are in reading_order?
+                if (all(subregion.id in reading_order for subregion in subregions) and
+                    isinstance(reading_order[subregions[0].id], # all have .index?
+                               (OrderedGroupType, OrderedGroupIndexedType))):
+                    subregions = sorted(subregions, key=lambda subregion:
+                                        reading_order[subregion.id].index)
+                region_unicode = page_element_unicode0(subregions[0])
+                for subregion, next_subregion in zip(subregions, subregions[1:]):
+                    if (subregion.id, next_subregion.id) not in joins:
+                        region_unicode += '\n' # or '\f'?
+                    region_unicode += page_element_unicode0(next_subregion)
+                region_conf = sum(page_element_conf0(subregion) for subregion in subregions)
+                region_conf /= len(subregions)
+            else: # TODO: what if a TextRegion has both TextLine and TextRegion children?
+                lines = region.get_TextLine()
+                if ((region.get_textLineOrder() or
+                     page.get_textLineOrder()) ==
+                    TextLineOrderSimpleType.BOTTOMTOTOP):
+                    lines = list(reversed(lines))
+                if level != 'line':
+                    for line in lines:
+                        words = line.get_Word()
+                        if ((line.get_readingDirection() or
+                             region.get_readingDirection() or
+                             page.get_readingDirection()) ==
+                            ReadingDirectionSimpleType.RIGHTTOLEFT):
+                            words = list(reversed(words))
+                        if level != 'word':
+                            for word in words:
+                                glyphs = word.get_Glyph()
+                                if ((word.get_readingDirection() or
+                                     line.get_readingDirection() or
+                                     region.get_readingDirection() or
+                                     page.get_readingDirection()) ==
+                                    ReadingDirectionSimpleType.RIGHTTOLEFT):
+                                    glyphs = list(reversed(glyphs))
+                                word_unicode = ''.join(page_element_unicode0(glyph) for glyph in glyphs)
+                                word_conf = sum(page_element_conf0(glyph) for glyph in glyphs)
+                                if glyphs:
+                                    word_conf /= len(glyphs)
+                                if not word.get_TextEquiv() or overwrite:
+                                    word.set_TextEquiv( # replace old, if any
+                                        [TextEquivType(Unicode=word_unicode, conf=word_conf)])
+                        line_unicode = ' '.join(page_element_unicode0(word) for word in words)
+                        line_conf = sum(page_element_conf0(word) for word in words)
+                        if words:
+                            line_conf /= len(words)
+                        if not line.get_TextEquiv() or overwrite:
+                            line.set_TextEquiv( # replace old, if any
+                                [TextEquivType(Unicode=line_unicode, conf=line_conf)])
+                region_unicode = ''
+                region_conf = 0
+                if lines:
+                    region_unicode = page_element_unicode0(lines[0])
+                    for line, next_line in zip(lines, lines[1:]):
+                        words = line.get_Word()
+                        next_words = next_line.get_Word()
+                        if not (words and next_words and (words[-1].id, next_words[0].id) in joins):
+                            region_unicode += '\n'
+                        region_unicode += page_element_unicode0(next_line)
+                    region_conf = sum(page_element_conf0(line) for line in lines)
+                    region_conf /= len(lines)
+            if not region.get_TextEquiv() or overwrite:
+                region.set_TextEquiv( # replace old, if any
+                    [TextEquivType(Unicode=region_unicode, conf=region_conf)])
+
+def page_shrink_higher_coordinate_levels(maxlevel, minlevel, pcgts):
+    """Project the coordinate hull of all PAGE-XML hierarchy levels above ``minlevel`` up to ``maxlevel``.
+    
+    Starting with the lowest hierarchy level chosen for processing,
+    join all segments into a convex hull for the next higher level,
+    replacing the parent coordinates, respectively.
+    
+    Follow regions recursively, but make sure to traverse them in a depth-first strategy.
+    """
+    LOG = getLogger('processor.TesserocrRecognize')
+    page = pcgts.get_Page()
+    regions = page.get_AllRegions(classes=['Text'])
+    if minlevel != 'region':
+        for region in regions:
+            lines = region.get_TextLine()
+            if minlevel != 'line':
+                for line in lines:
+                    words = line.get_Word()
+                    if minlevel != 'word':
+                        for word in words:
+                            glyphs = word.get_Glyph()
+                            if maxlevel in ['region', 'line', 'word', 'glyph'] and glyphs:
+                                joint_polygon = join_segments(glyphs)
+                                LOG.debug("setting hull for word '%s' from %d vertices",
+                                          word.id, len(joint_polygon))
+                                word.get_Coords().set_points(points_from_polygon(joint_polygon))
+                    if maxlevel in ['region', 'line', 'word'] and words:
+                        joint_polygon = join_segments(words)
+                        LOG.debug("setting hull for line '%s' from %d vertices",
+                                  line.id, len(joint_polygon))
+                        line.get_Coords().set_points(points_from_polygon(joint_polygon))
+            if maxlevel in ['region', 'line'] and lines:
+                joint_polygon = join_segments(lines)
+                LOG.debug("setting hull for region '%s' from %d vertices",
+                          region.id, len(joint_polygon))
+                region.get_Coords().set_points(points_from_polygon(joint_polygon))
+
+def join_segments(segments):
+    return join_polygons([polygon_from_points(segment.get_Coords().points)
+                          for segment in segments])
+
+def join_polygons(polygons, scale=20):
+    """construct concave hull (alpha shape) from input polygons by connecting their pairwise nearest points"""
+    return make_join([make_valid(Polygon(poly)) for poly in polygons], scale=scale).exterior.coords[:-1]
+
+def make_join(polygons, scale=20):
+    """construct concave hull (alpha shape) from input polygons by connecting their pairwise nearest points"""
+    # ensure input polygons are simply typed and all oriented equally
+    polygons = [orient(poly)
+                for poly in itertools.chain.from_iterable(
+                        [poly.geoms
+                         if poly.geom_type in ['MultiPolygon', 'GeometryCollection']
+                         else [poly]
+                         for poly in polygons])]
+    npoly = len(polygons)
+    if npoly == 1:
+        return polygons[0]
+    # find min-dist path through all polygons (travelling salesman)
+    pairs = itertools.combinations(range(npoly), 2)
+    dists = np.zeros((npoly, npoly), dtype=float)
+    for i, j in pairs:
+        dist = polygons[i].distance(polygons[j])
+        if dist < 1e-5:
+            dist = 1e-5 # if pair merely touches, we still need to get an edge
+        dists[i, j] = dist
+        dists[j, i] = dist
+    dists = minimum_spanning_tree(dists, overwrite=True)
+    # add bridge polygons (where necessary)
+    for prevp, nextp in zip(*dists.nonzero()):
+        prevp = polygons[prevp]
+        nextp = polygons[nextp]
+        nearest = nearest_points(prevp, nextp)
+        bridgep = LineString(nearest).buffer(max(1, scale/5), resolution=1)
+        polygons.append(bridgep)
+    jointp = unary_union(polygons)
+    assert jointp.geom_type == 'Polygon', jointp.wkt
+    # follow-up calculations will necessarily be integer;
+    # so anticipate rounding here and then ensure validity
+    jointp2 = set_precision(jointp, 1.0)
+    if jointp2.geom_type != 'Polygon' or not jointp2.is_valid:
+        jointp2 = Polygon(np.round(jointp.exterior.coords))
+        jointp2 = make_valid(jointp2)
+    assert jointp2.geom_type == 'Polygon', jointp2.wkt
+    return jointp2
+
+def pad_image(image, padding):
+    # TODO: input padding can create extra edges if not binarized; at least try to smooth
+    stat = ImageStat.Stat(image)
+    # workaround for Pillow#4925
+    if len(stat.bands) > 1:
+        background = tuple(stat.median)
+    else:
+        background = stat.median[0]
+    padded = Image.new(image.mode,
+                       (image.width + 2 * padding,
+                        image.height + 2 * padding),
+                       background)
+    padded.paste(image, (padding, padding))
+    return padded
+
+def polygon_for_parent(polygon, parent):
+    """Clip polygon to parent polygon range.
+    
+    (Should be moved to ocrd_utils.coordinates_for_segment.)
+    """
+    childp = Polygon(polygon)
+    if isinstance(parent, PageType):
+        if parent.get_Border():
+            parentp = Polygon(polygon_from_points(parent.get_Border().get_Coords().points))
+        else:
+            parentp = Polygon([[0, 0], [0, parent.get_imageHeight()],
+                               [parent.get_imageWidth(), parent.get_imageHeight()],
+                               [parent.get_imageWidth(), 0]])
+    else:
+        parentp = Polygon(polygon_from_points(parent.get_Coords().points))
+    # ensure input coords have valid paths (without self-intersection)
+    # (this can happen when shapes valid in floating point are rounded)
+    childp = make_valid(childp)
+    parentp = make_valid(parentp)
+    if not childp.is_valid:
+        return None
+    if not parentp.is_valid:
+        return None
+    # check if clipping is necessary
+    if childp.within(parentp):
+        return childp.exterior.coords[:-1]
+    # clip to parent
+    interp = make_intersection(childp, parentp)
+    if not interp:
+        return None
+    return interp.exterior.coords[:-1] # keep open
+
+def make_intersection(poly1, poly2):
+    interp = poly1.intersection(poly2)
+    # post-process
+    if interp.is_empty or interp.area == 0.0:
+        # this happens if Tesseract "finds" something
+        # outside of the valid Border of a deskewed/cropped page
+        # (empty corners created by masking); will be ignored
+        return None
+    if interp.geom_type == 'GeometryCollection':
+        # heterogeneous result: filter zero-area shapes (LineString, Point)
+        interp = unary_union([geom for geom in interp.geoms if geom.area > 0])
+    if interp.geom_type == 'MultiPolygon':
+        # homogeneous result: construct convex hull to connect
+        interp = make_join(interp.geoms)
+    if interp.minimum_clearance < 1.0:
+        # follow-up calculations will necessarily be integer;
+        # so anticipate rounding here and then ensure validity
+        interp = Polygon(np.round(interp.exterior.coords))
+        interp = make_valid(interp)
+    return interp
+
+def make_valid(polygon):
+    points = list(polygon.exterior.coords)
+    for split in range(1, len(points)):
+        if polygon.is_valid or polygon.simplify(polygon.area).is_valid:
+            break
+        # simplification may not be possible (at all) due to ordering
+        # in that case, try another starting point
+        polygon = Polygon(points[-split:]+points[:-split])
+    # try by simplification
+    for tolerance in range(int(polygon.area + 1.5)):
+        if polygon.is_valid:
+            break
+        # simplification may require a larger tolerance
+        polygon = polygon.simplify(tolerance + 1)
+    # try by enlarging
+    for tolerance in range(1, int(polygon.area + 2.5)):
+        if polygon.is_valid:
+            break
+        # enlargement may require a larger tolerance
+        polygon = polygon.buffer(tolerance)
+    assert polygon.is_valid, polygon.wkt
+    return polygon
+
+def iterate_level(it, ril, parent=None):
+    LOG = getLogger('processor.TesserocrRecognize')
+    # improves over tesserocr.iterate_level by
+    # honouring multi-level semantics so iterators
+    # can be combined across levels
+    if parent is None:
+        parent = ril - 1
+    pos = 0
+    while it and not it.Empty(ril):
+        yield it
+        # With upstream Tesseract, these assertions may fail:
+        # if ril > 0 and it.IsAtFinalElement(parent, ril):
+        #     for level in range(parent, ril):
+        #         assert it.IsAtFinalElement(parent, level), \
+        #             "level %d iterator at %d is final w.r.t. %d but level %d is not" % (
+        #                 ril, pos, parent, level)
+        # Hence the following workaround avails itself:
+        if ril > 0 and all(it.IsAtFinalElement(parent, level)
+                           for level in range(parent, ril + 1)):
+            break
+        if not it.Next(ril):
+            break
+        while it.Empty(ril) and not it.Empty(0):
+            # This happens when
+            # - on RIL.PARA, RIL.TEXTLINE and RIL.WORD,
+            #   empty non-text (pseudo-) blocks intervene
+            # - on RIL.SYMBOL, a word has no cblobs at all
+            #   (because they have all been rejected)
+            # We must _not_ yield these (as they have strange
+            # properties and bboxes). But most importantly,
+            # they will have met IsAtFinalElement prematurely
+            # (hence the similar loop above).
+            # Since this may happen multiple consecutive times,
+            # enclose this in a while loop.
+            LOG.warning("level %d iterator at %d needs to skip empty segment",
+                        ril, pos)
+            if not it.Next(ril):
+                break
+        pos += 1
