+++ ocrd_tesserocr/crop.py
@@ -2,8 +2,8 @@ from __future__ import absolute_import
 import os.path
 
 import tesserocr
+
 from ocrd_utils import (
-    getLogger,
     crop_image,
     coordinates_for_segment,
     coordinates_of_segment,
@@ -12,29 +12,34 @@ from ocrd_utils import (
     polygon_from_bbox,
     points_from_polygon,
     bbox_from_xywh,
-    make_file_id,
-    assert_file_grp_cardinality,
-    MIMETYPE_PAGE
 )
-from ocrd_modelfactory import page_from_file
 from ocrd_models.ocrd_page import (
     CoordsType,
     AlternativeImageType,
     BorderType,
-    to_xml
 )
 
-from .recognize import TesserocrRecognize, polygon_for_parent
+from .recognize import TesserocrRecognize
+from .common import polygon_for_parent
 
 class TesserocrCrop(TesserocrRecognize):
     @property
     def executable(self):
         return 'ocrd-tesserocr-crop'
 
-    def process(self):
+    def _init(self):
+        # use default model (eng) with vanilla tesserocr API
+        self.tessapi = tesserocr.PyTessBaseAPI()
+        # disable table detection here (tables count as text blocks),
+        # because we do not want to risk confusing the spine with
+        # a column separator and thus creeping into a neighbouring
+        # page:
+        self.tessapi.SetVariable("textord_tabfind_find_tables", "0")
+
+    def process_page_pcgts(self, pcgts, output_file_id=None, page_id=None):
         """Performs page cropping with Tesseract on the workspace.
 
-        Open and deserialize PAGE input files and their respective images.
+        Open and deserialize PAGE input file and its respective images.
         Set up Tesseract to detect text blocks on each page, and find
         the largest coordinate extent spanning all of them. Use this
         extent in defining a Border, and add that to the page.
@@ -48,81 +53,59 @@ class TesserocrCrop(TesserocrRecognize):
         
         Produce new output files by serialising the resulting hierarchy.
         """
-        assert_file_grp_cardinality(self.input_file_grp, 1)
-        assert_file_grp_cardinality(self.output_file_grp, 1)
-
-        with tesserocr.PyTessBaseAPI() as tessapi:
-            # disable table detection here (tables count as text blocks),
-            # because we do not want to risk confusing the spine with
-            # a column separator and thus creeping into a neighbouring
-            # page:
-            tessapi.SetVariable("textord_tabfind_find_tables", "0")
-            for (n, input_file) in enumerate(self.input_files):
-                file_id = make_file_id(input_file, self.output_file_grp)
-                page_id = input_file.pageId or input_file.ID
-                self.logger.info("INPUT FILE %i / %s", n, page_id)
-                pcgts = page_from_file(self.workspace.download_file(input_file))
-                self.add_metadata(pcgts)
-                page = pcgts.get_Page()
+        page = pcgts.get_Page()
                 
-                # warn of existing Border:
-                border = page.get_Border()
-                if border:
-                    left, top, right, bottom = bbox_from_points(border.get_Coords().points)
-                    self.logger.warning('Overwriting existing Border: %i:%i,%i:%i',
-                                        left, top, right, bottom)
-                
-                page_image, page_xywh, page_image_info = self.workspace.image_from_page(
-                    page, page_id,
-                    # image must not have been cropped already,
-                    # abort if no such image can be produced:
-                    feature_filter='cropped')
-                if self.parameter['dpi'] > 0:
-                    dpi = self.parameter['dpi']
-                    self.logger.info("Page '%s' images will use %d DPI from parameter override", page_id, dpi)
-                elif page_image_info.resolution != 1:
-                    dpi = page_image_info.resolution
-                    if page_image_info.resolutionUnit == 'cm':
-                        dpi = round(dpi * 2.54)
-                    self.logger.info("Page '%s' images will use %d DPI from image meta-data", page_id, dpi)
-                else:
-                    dpi = 0
-                    self.logger.info("Page '%s' images will use DPI estimated from segmentation", page_id)
-                tessapi.SetVariable('user_defined_dpi', str(dpi))
-                if dpi:
-                    zoom = 300 / dpi
-                else:
-                    zoom = 1
+        # warn of existing Border:
+        border = page.get_Border()
+        if border:
+            left, top, right, bottom = bbox_from_points(border.get_Coords().points)
+            self.logger.warning('Overwriting existing Border: %i:%i,%i:%i',
+                                left, top, right, bottom)
 
-                bounds = self.estimate_bounds(page, page_image, tessapi, zoom)
-                self.process_page(page, page_image, page_xywh, bounds, file_id, input_file.pageId)
+        page_image, page_xywh, page_image_info = self.workspace.image_from_page(
+            page, page_id,
+            # image must not have been cropped already,
+            # abort if no such image can be produced:
+            feature_filter='cropped')
+        if self.parameter['dpi'] > 0:
+            dpi = self.parameter['dpi']
+            self.logger.info("Page '%s' images will use %d DPI from parameter override", page_id, dpi)
+        elif page_image_info.resolution != 1:
+            dpi = page_image_info.resolution
+            if page_image_info.resolutionUnit == 'cm':
+                dpi = round(dpi * 2.54)
+            self.logger.info("Page '%s' images will use %d DPI from image meta-data", page_id, dpi)
+        else:
+            dpi = 0
+            self.logger.info("Page '%s' images will use DPI estimated from segmentation", page_id)
+        self.tessapi.SetVariable('user_defined_dpi', str(dpi))
+        if dpi:
+            zoom = 300 / dpi
+        else:
+            zoom = 1
 
-                pcgts.set_pcGtsId(file_id)
-                self.workspace.add_file(
-                    file_id=file_id,
-                    file_grp=self.output_file_grp,
-                    page_id=input_file.pageId,
-                    mimetype=MIMETYPE_PAGE,
-                    local_filename=os.path.join(self.output_file_grp,
-                                                file_id + '.xml'),
-                    content=to_xml(pcgts))
+        bounds = self._estimate_bounds(page, page_image, zoom)
+        cropped = self._process_page(page, page_image, page_xywh, bounds, output_file_id)
+        if cropped:
+            return [pcgts, cropped]
+        return pcgts
         
-    def estimate_bounds(self, page, page_image, tessapi, zoom=1.0):
+    def _estimate_bounds(self, page, page_image, zoom=1.0):
         """Get outer bounds of all (existing or detected) regions."""
         all_left = page_image.width
         all_top = page_image.height
         all_right = 0
         all_bottom = 0
         self.logger.info("Cropping with Tesseract")
-        tessapi.SetImage(page_image)
+        self.tessapi.SetImage(page_image)
         # PSM.SPARSE_TEXT: get as much text as possible in no particular order
         # PSM.AUTO (default): includes tables (dangerous)
         # PSM.SPARSE_TEXT_OSD: sparse but all orientations
-        tessapi.SetPageSegMode(tesserocr.PSM.SPARSE_TEXT)
+        self.tessapi.SetPageSegMode(tesserocr.PSM.SPARSE_TEXT)
         #
         # iterate over all text blocks and compare their
         # bbox extent to the running min and max values
-        for component in tessapi.GetComponentImages(tesserocr.RIL.BLOCK, True):
+        for component in self.tessapi.GetComponentImages(tesserocr.RIL.BLOCK, True):
             image, xywh, index, _ = component
             #
             # the region reference in the reading order element
@@ -165,12 +148,12 @@ class TesserocrCrop(TesserocrRecognize):
                          all_left, all_right, all_top, all_bottom)
         return all_left, all_top, all_right, all_bottom
 
-    def process_page(self, page, page_image, page_xywh, bounds, file_id, page_id):
+    def _process_page(self, page, page_image, page_xywh, bounds, file_id):
         """Set the identified page border, if valid."""
         left, top, right, bottom = bounds
         if left >= right or top >= bottom:
-            self.logger.error("Cannot find valid extent for page '%s'", page_id)
-            return
+            self.logger.error("Cannot find valid extent for page")
+            return False
         padding = self.parameter['padding']
         # add padding:
         left = max(left - padding, 0)
@@ -183,7 +166,7 @@ class TesserocrCrop(TesserocrRecognize):
         polygon = polygon_for_parent(polygon, page)
         if polygon is None:
             self.logger.error("Ignoring extant border")
-            return
+            return False
         border = BorderType(Coords=CoordsType(
             points_from_polygon(polygon)))
         # intersection with parent could have changed bbox,
@@ -194,10 +177,9 @@ class TesserocrCrop(TesserocrRecognize):
         # update METS (add the image file):
         page_image = crop_image(page_image, box=bbox)
         page_xywh['features'] += ',cropped'
-        file_path = self.workspace.save_image_file(
-            page_image, file_id + '.IMG-CROP',
-            page_id=page_id, file_grp=self.output_file_grp)
+        page_image_id = file_id + '.IMG-CROP'
+        page_image_path = os.path.join(self.output_file_grp, page_image_id + '.png')
         # update PAGE (reference the image file):
         page.add_AlternativeImage(AlternativeImageType(
-            filename=file_path, comments=page_xywh['features']))
-    
+            filename=page_image_path, comments=page_xywh['features']))
+        return page_image, page_image_id, page_image_path
