+++ ocrd_tesserocr/segment_line.py
@@ -20,10 +20,10 @@ class TesserocrSegmentLine(TesserocrRecognize):
             # add default params
             assert ParameterValidator(self.metadata['tools']['ocrd-tesserocr-recognize']).validate(self.parameter).is_valid
 
-    def process(self):
+    def process_page_pcgts(self, pcgts, output_file_id=None, page_id=None):
         """Performs (text) line segmentation with Tesseract on the workspace.
         
-        Open and deserialize PAGE input files and their respective images,
+        Open and deserialize PAGE input file and its respective images,
         then iterate over the element hierarchy down to the (text) region level,
         and remove any existing TextLine elements (unless ``overwrite_lines``
         is False).
@@ -39,4 +39,4 @@ class TesserocrSegmentLine(TesserocrRecognize):
         
         Produce a new output file by serialising the resulting hierarchy.
         """
-        super(TesserocrSegmentLine, self).process()
+        return super().process_page_pcgts(pcgts, output_file_id=output_file_id, page_id=page_id)
